+++ scripts/prompt_enhance.py
@@ -95,5 +95,5 @@ class Script(scripts.Script):
             shared.log.debug(f'Prompt enhance: prompt="{p.prompt}"')
 
     def after_component(self, component, **kwargs): # searching for actual ui prompt components
-        if getattr(component, 'elem_id', '') in ['txt2img_prompt', 'img2img_prompt', 'control_prompt']:
+        if getattr(component, 'elem_id', '') in ['txt2img_prompt', 'img2img_prompt', 'control_prompt', 'video_prompt']:
             self.prompt = component